TransactionTracers are now SpanTracers
